+++ src/Models/Transactions/AuthorizeTransactionRequest.cs
@@ -12,9 +12,9 @@ namespace epay3.Sdk.Models
         public string TokenId { get; set; }
 
         /// <summary>
-        /// Gets or sets the total amount to authorize.
+   /// Gets or sets the total amount to authorize.
         /// Required.
         /// </summary>
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 }
